[refactor] パッケージ構造を単純化
